Emotion_Recognition_Pytorch/requirements.txt
attrs==21.2.0
blis==0.7.5
cached-property==1.5.2
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cmake==3.23.3
contextvars==2.4
cycler==0.11.0
cymem==2.0.6
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
dlib==19.24.0
enum34==1.1.10
future==0.18.2
googledrivedownloader==0.4
h5py==3.1.0
idna==3.3
imageio==2.15.0
immutables==0.16
importlib-metadata==4.8.2
iniconfig==1.1.1
isodate==0.6.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.1
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib==3.3.4
murmurhash==1.0.6
networkx==2.5.1
nltk==3.6.5
numpy==1.19.5
opencv-python==4.6.0.66
ordered-set==4.0.2
packaging==21.3
pandas==1.1.5
pathy==0.6.1
Pillow==8.4.0
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
pydantic==1.8.2
pynvml==11.4.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
rdflib==5.0.0
regex==2021.11.10
requests==2.26.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tifffile==2020.9.3
toml==0.10.2
torch==1.9.1+cu111
torch-geometric==2.0.2
torch-scatter==2.0.9
torch-sparse==0.6.12
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.7.4.3
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
wrapt==1.12.1
yacs==0.1.8
zipp==3.6.0
